Move files to feature
