+++ packages/core/src/features/population-of-logs-to-a-file/common/ipc-file-logger-channel.ts
