+++ packages/core/src/features/population-of-logs-to-a-file/main/create-ipc-file-transport.global-override-for-injectable.ts
