+++ packages/core/src/features/population-of-logs-to-a-file/main/ipc-file-logger.injectable.ts
