+++ packages/core/src/features/population-of-logs-to-a-file/main/ipc-file-logger.test.ts
@@ -2,7 +2,7 @@
  * Copyright (c) OpenLens Authors. All rights reserved.
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
-import { getDiForUnitTesting } from "../getDiForUnitTesting";
+import { getDiForUnitTesting } from "../../../main/getDiForUnitTesting";
 import createIpcFileLoggerTransportInjectable from "./create-ipc-file-transport.injectable";
 import type { IpcFileLogger } from "./ipc-file-logger.injectable";
 import ipcFileLoggerInjectable from "./ipc-file-logger.injectable";
