+++ packages/core/src/features/population-of-logs-to-a-file/main/stop-ipc-logging.injectable.ts
@@ -3,7 +3,7 @@
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
 import { getInjectable } from "@ogre-tools/injectable";
-import { beforeQuitOfFrontEndInjectionToken } from "../start-main-application/runnable-tokens/phases";
+import { beforeQuitOfFrontEndInjectionToken } from "../../../main/start-main-application/runnable-tokens/phases";
 import ipcFileLoggerInjectable from "./ipc-file-logger.injectable";
 
 const stopIpcLoggingInjectable = getInjectable({
