+++ packages/core/src/features/population-of-logs-to-a-file/renderer/close-renderer-log-file-id.test.ts
@@ -6,8 +6,8 @@ import type winston from "winston";
 import type { SendMessageToChannel } from "@k8slens/messaging";
 import { sendMessageToChannelInjectionToken } from "@k8slens/messaging";
 import type { DiContainer } from "@ogre-tools/injectable";
-import winstonLoggerInjectable from "../../common/winston-logger.injectable";
-import { getDiForUnitTesting } from "../getDiForUnitTesting";
+import winstonLoggerInjectable from "../../../common/winston-logger.injectable";
+import { getDiForUnitTesting } from "../../../renderer/getDiForUnitTesting";
 import closeRendererLogFileInjectable from "./close-renderer-log-file.injectable";
 import rendererLogFileIdInjectable from "./renderer-log-file-id.injectable";
 import ipcLogTransportInjectable from "./ipc-transport.injectable";
