+++ packages/core/src/features/population-of-logs-to-a-file/renderer/close-renderer-log-file.injectable.ts
@@ -4,8 +4,8 @@
  */
 import { sendMessageToChannelInjectionToken } from "@k8slens/messaging";
 import { getInjectable } from "@ogre-tools/injectable";
-import winstonLoggerInjectable from "../../common/winston-logger.injectable";
-import { closeIpcFileLoggerChannel } from "../../common/logger/ipc-file-logger-channel";
+import winstonLoggerInjectable from "../../../common/winston-logger.injectable";
+import { closeIpcFileLoggerChannel } from "../common/ipc-file-logger-channel";
 import rendererLogFileIdInjectable from "./renderer-log-file-id.injectable";
 import ipcLogTransportInjectable from "./ipc-transport.injectable";
