+++ packages/core/src/features/population-of-logs-to-a-file/renderer/ipc-transport.ts
