+++ packages/core/src/renderer/before-frame-starts/runnables/listen-unload.injectable.ts
@@ -2,14 +2,14 @@
  * Copyright (c) OpenLens Authors. All rights reserved.
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
+import { unmountComponentAtNode } from "react-dom";
 import { getInjectable } from "@ogre-tools/injectable";
 import currentlyInClusterFrameInjectable from "../../routes/currently-in-cluster-frame.injectable";
 import { beforeFrameStartsSecondInjectionToken } from "../tokens";
 import loggerInjectable from "../../../common/logger.injectable";
 import hostedClusterInjectable from "../../cluster-frame-context/hosted-cluster.injectable";
 import frameRoutingIdInjectable from "../../frames/cluster-frame/init-cluster-frame/frame-routing-id/frame-routing-id.injectable";
-import closeRendererLogFileInjectable from "../../logger/close-renderer-log-file.injectable";
-import { unmountComponentAtNode } from "react-dom";
+import closeRendererLogFileInjectable from "../../../features/population-of-logs-to-a-file/renderer/close-renderer-log-file.injectable";
 
 const listenUnloadInjectable = getInjectable({
   id: "listen-unload",
